+++ engine/apps/oss_installation/models/cloud_connector.py
@@ -69,7 +69,7 @@ class CloudConnector(models.Model):
         page = 1
         while fetch_next_page:
             try:
-                url = urljoin(users_url, f"?page={page}&?short=true")
+                url = urljoin(users_url, f"?page={page}&short=true&roles=0&roles=1")
                 r = requests.get(url, headers={"AUTHORIZATION": api_token}, timeout=5)
                 if r.status_code != 200:
                     logger.warning(
@@ -115,7 +115,7 @@ class CloudConnector(models.Model):
             logger.warning(f"Unable to sync_user_with cloud user_id {user.id}. GRAFANA_CLOUD_ONCALL_TOKEN is not set")
             error_msg = "GRAFANA_CLOUD_ONCALL_TOKEN is not set"
         else:
-            url = urljoin(GRAFANA_CLOUD_ONCALL_API_URL, f"api/v1/users/?email={user.email}")
+            url = urljoin(GRAFANA_CLOUD_ONCALL_API_URL, f"api/v1/users/?email={user.email}&roles=0&roles=1&short=true")
             try:
                 r = requests.get(url, headers={"AUTHORIZATION": api_token}, timeout=5)
                 if r.status_code != 200: