+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 1.0.4 (2022-06-28)
+- Allow Telegram DMs without channel connection.
+
+## 1.0.3 (2022-06-27)
+- Fix users public api endpoint. Now it returns users with all roles.
+- Fix redundant notifications about gaps in schedules.
+- Frontend fixes.
+
+## 1.0.2 (2022-06-17)
+
+- Fix Grafana Alerting integration to handle API changes in Grafana 9
+- Improve public api endpoint for for outgoing webhooks (/actions) by adding ability to create, update and delete outgoing webhook instance
+
+## 1.0.0 (2022-06-14)
+
+- First Public Release
+
 ## 0.0.71 (2022-06-06)
 
-- Initial Release
+- Initial Commit Release